+++ packages/monte_carlo/collision/photon/src/MonteCarlo_PhotoelectricPhotoatomicReaction.hpp
@@ -31,13 +31,13 @@ public:
   PhotoelectricPhotoatomicReaction(
        const std::shared_ptr<const std::vector<double> >& incoming_energy_grid,
        const std::shared_ptr<const std::vector<double> >& cross_section,
-       const unsigned threshold_energy_index );
+       const size_t threshold_energy_index );
 
   //! Constructor
   PhotoelectricPhotoatomicReaction(
      const std::shared_ptr<const std::vector<double> >& incoming_energy_grid,
      const std::shared_ptr<const std::vector<double> >& cross_section,
-     const unsigned threshold_energy_index,
+     const size_t threshold_energy_index,
      const std::shared_ptr<const Utility::HashBasedGridSearcher<double> >&
      grid_searcher );
